refactor: Date 객체 관련 유틸 함수 구현 및 적용
