+++ src/utils/date.ts
@@ -1,0 +1,9 @@
+export const dateToYMDHMString = (date: Date) => {
+  const year = date.getFullYear();
+  const month = (date.getMonth() + 1).toString().padStart(2, "0");
+  const day = date.getDate().toString().padStart(2, "0");
+  const hour = date.getHours();
+  const min = date.getMinutes().toString().padStart(2, "0");
+
+  return `${year}-${month}-${day} ${hour}:${min}`;
+};
